+++ src/filter/LowShelfFilter.cpp
@@ -24,7 +24,7 @@ dibiff::filter::LowShelfFilter::LowShelfFilter(float gain, float cutoff, float s
  * @param qFactor The quality factor of the filter
  */
 dibiff::filter::Coefficients dibiff::filter::LowShelfFilter::calculateCoefficients(float gain, float cutoff, float sampleRate, float qFactor) {
-    float A = std::powf(10.0f, gain / 40.0f);
+    float A = std::pow(10.0f, gain / 40.0f);
     float w0 = 2.0f * M_PI * cutoff / sampleRate;
     float alpha = std::sinf(w0) / (2.0f * qFactor);
     float cosw0 = std::cosf(w0);
